% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hd}
\alias{hd}
\title{Compute the Harrell-Davis estimate of the qth quantile}
\usage{
hd(x, q, cores = 1L)
}
\arguments{
\item{x}{a numeric \code{vector}}

\item{q}{a desired quantile}

\item{cores}{a number of cores used}
}
\value{
the Harrell-Davis estimate of the \code{q}-th quantile of \code{x}
}
\description{
Compute the Harrell-Davis estimate of the qth quantile
}

