% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hdci}
\alias{hdci}
\title{Compute a standard error of the Harrell-Davis of the qth quantile}
\usage{
hdci(x, q = 0.5, alpha = 0.05, nboot = 100, SEED = TRUE, pr = TRUE,
  cores = 1)
}
\arguments{
\item{x}{a numeric \code{vector}}

\item{q}{a desired quantile}

\item{alpha}{a significance level}

\item{nboot}{a number of bootstraps}

\item{SEED}{}

\item{pr}{}
}
\description{
Compute a standard error of the Harrell-Davis of the qth quantile
}
\author{
Rand R. Wilcox

Masahiro Kanai
}

