% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hdpb}
\alias{hdpb}
\title{Compute a bootstrap confidence interval for the Harrell-Davis estimate of the qth quantile}
\usage{
hdpb(x, q, alpha = 0.05, nboot = 2000L, nv = 0, cores = 1L)
}
\arguments{
\item{x}{a numeric \code{vector}}

\item{q}{a desired quantile}

\item{alpha}{a significance level}

\item{nboot}{a number of bootstraps}

\item{nv}{a null-value when computing a p-value}

\item{cores}{a number of cores used}
}
\value{
the confidence inteval
}
\description{
Compute a bootstrap confidence interval for the Harrell-Davis estimate of the qth quantile
}

