% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hdseb}
\alias{hdseb}
\title{Compute a bootstrap standard error of the Harrell-Davis estimate of the qth quantile}
\usage{
hdseb(x, q = 0.5, nboot = 100, SEED = TRUE, cores = 1)
}
\arguments{
\item{x}{a numeric \code{vector}}

\item{q}{a desired quantile}

\item{nboot}{a number of bootstraps}

\item{SEED}{a logical indicating whether the seed should be set}

\item{cores}{a number of cores used for computation}
}
\value{
the standard error of the Harrell-Davis estimate of the qth quantile
}
\description{
Compute a bootstrap standard error of the Harrell-Davis estimate of the qth quantile
}

