% Generated by roxygen2 (4.0.2): do not edit by hand
\name{qcomhd}
\alias{qcomhd}
\title{Compare quantiles via the Harrell-Davis estimator}
\usage{
qcomhd(x, y, q = c(0.1, 0.25, 0.5, 0.75, 0.9), nboot = 2000,
  plotit = TRUE, SEED = TRUE, xlab = "Group 1", ylab = "Est.1-Est.2",
  alpha = 0.05, cores = 1)
}
\arguments{
\item{x}{a numeric \code{vector}}

\item{y}{a numeric \code{vector}}

\item{q}{a desired quantile}

\item{nboot}{a number of bootstraps}

\item{plotit}{}

\item{SEED}{}

\item{xlab}{}

\item{ylab}{}

\item{alpha}{a significance level}
}
\description{
Compare quantiles via the Harrell-Davis estimator
}

