% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{WRShd}
\alias{WRShd}
\alias{WRShd-package}
\title{WRShd}
\description{
An Experimental Refactor of the Harrell-Davis Estimate Functions from R. R. Wilcox' Robust Statistics Package "WRS"
}
\author{
Masahiro Kanai

Rand R. Wilcox

Felix Schönbrodt
}

