% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hd}
\alias{hd}
\title{Compute the Harrell-Davis estimate of the qth quantile}
\usage{
hd(x, q = 0.5, na.rm = TRUE, cores = 1)
}
\arguments{
\item{x}{a numeric \code{vector}}

\item{q}{a desired quantile}

\item{na.rm}{a logical indicating whether missing values should be removed (should be TRUE)}

\item{cores}{a number of cores used for computation}
}
\value{
the Harrell-Davis estimate of the qth quantile of \code{x}
}
\description{
Compute the Harrell-Davis estimate of the qth quantile
}

