% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hdci}
\alias{hdci}
\title{Compute a 1-\code{alpha} confidence for the Harrell-Davis estimate of the qth quantile}
\usage{
hdci(x, q = 0.5, alpha = 0.05, nboot = 100, SEED = TRUE, pr = TRUE,
  cores = 1)
}
\arguments{
\item{x}{a numeric \code{vector}}

\item{q}{a desired quantile}

\item{alpha}{a significance level}

\item{nboot}{a number of bootstraps}

\item{SEED}{a logical indicating whether the seed should be set}

\item{pr}{a logical indicating whether}

\item{cores}{a number of cores used for computation}
}
\value{
the 1-\code{alpha} confidence for the Harrell-Davis estimate of the qth quantile
}
\description{
Compute a 1-\code{alpha} confidence for the Harrell-Davis estimate of the qth quantile
}

