% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hdpb}
\alias{hdpb}
\title{Compute a bootstrap 1-\code{alpha} confidence for the Harrell-Davis estimate of the qth quantile}
\usage{
hdpb(x, q = 0.5, alpha = 0.05, nboot = 2000, SEED = TRUE, nv = 0,
  cores = 1)
}
\arguments{
\item{x}{a numeric \code{vector}}

\item{q}{a desired quantile}

\item{alpha}{a significance level}

\item{nboot}{a number of bootstraps}

\item{SEED}{a logical indicating whether the seed should be set}

\item{nv}{a null-value when computing a p-value}

\item{cores}{a number of cores used for computation}
}
\value{
the a bootstrap 1-\code{alpha} confidence for the Harrell-Davis estimate of the qth quantile
}
\description{
Compute a bootstrap 1-\code{alpha} confidence for the Harrell-Davis estimate of the qth quantile
}

