% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hdseb}
\alias{hdseb}
\title{Compute a bootstrap standard error of the Harrell-Davis estimate of the qth quantile}
\usage{
hdseb(x, q = 0.5, nboot = 100L, cores = 1L)
}
\arguments{
\item{x}{a numeric \code{vector}}

\item{q}{a desired quantile}

\item{nboot}{a number of bootstraps}

\item{cores}{a number of cores used}
}
\value{
the standard error of the Harrell-Davis estimate of the qth quantile
}
\description{
Compute a bootstrap standard error of the Harrell-Davis estimate of the qth quantile
}

